+++ src/EscapeFromDarkness.cpp
@@ -1,0 +1,15 @@
+#include "EscapeFromDarkness.h"
+
+int main(int argc, char* argv[]) {
+    auto dungeon = new Dungeon(argc, argv);
+
+    while (dungeon->continueGame()) {
+        dungeon->getEventManager()->handleNext();
+        dungeon->getPlayer()->incrementDaysSurvived();
+    }
+    dungeon->getOutput()->printEndgame();
+
+    delete (dungeon);
+
+    return 0;
+}