+++ src/EscapeFromDarkness.h
@@ -1,11 +1,9 @@
 #ifndef ESCAPE_FROM_DARKNESS_H
 #define ESCAPE_FROM_DARKNESS_H
 
-#include <forward_declarations.h>
 #include <global.h>
 
-#include "dungeon.h"
-#include "characters/monster.h"
+#include "Dungeon.h"
 
 int game_tick(Dungeon* dungeon);